Add files via upload
Alzheimer’s disease (AD) is the most common form of dementia that can lead to a neurological brain disorder that causes progressive memory loss as a result of damaging the brain cells and the ability to perform daily activities. Using MRI (Magnetic Resonance Imaging) scan brain images, we can get the help of Artificial intelligence (AI) technology for detection and prediction of this disease and classify the AD patients. The main Objective of this project is to build the best prediction and detection model. In recent years, the Deep Learning (DL) algorithms are very useful for the diagnosis of AD as DL algorithms work well with large datasets. In this project, I have implemented Convolutional Neural Network (CNN) for the earlier diagnosis and classification of AD using MRI images.